Starting files
Starting files for the Simon game I'm trying to recreate. The files consist of sounds
which will be matched with a certain button pressed and the html index file which 
will serve as a canvas for the game. Attached is also the stylesheet which
will be used to style the web app.